requirements.txt
aniso8601==8.0.0
astroid==2.2.5
beautifulsoup4==4.8.2
blessings==1.7
bpython==0.18
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
curtsies==0.3.0
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
greenlet==0.4.15
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.0
idna==2.9
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.1
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pylint==2.3.1
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.23.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
soupsieve==2.0
spotipy==2.16.0
stem==1.8.0
torrequest==0.1.0
typed-ast==1.3.5
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.11.1
